DS-5299 ALOS Palsar More Info link updated
The link needs to be, "https://www.earthdata.nasa.gov/data/instruments/palsar" until an ASF URL exists for it. The current ASF URLs redirect to the earthdata site. A redirect in our doc modal causes problems, so the direct link to earthdata needs to be used until we have a native ASF url.